+++ insights-ui/docker-compose.yaml
@@ -1,0 +1,15 @@
+version: '3.1'
+
+services:
+  db:
+    image: postgres                # Uses the official PostgreSQL image
+    container_name: insights-ui-db     # Names the container for easy reference
+    restart: always                 # Automatically restarts the container if it stops
+    environment:
+      POSTGRES_USER: admin          # Sets the database user to 'admin'
+      POSTGRES_PASSWORD: admin      # Sets the database password to 'admin'
+      POSTGRES_DB: insights_ui_db  # Creates a database named 'next_app_localhost_db'
+    ports:
+      - "5432:5432"                 # Exposes port 5432 to the host
+    volumes:
+      - ./data:/var/lib/postgresql/data   # Persists data to the './data' folder
