--- insights-ui/src/app/public-equities/ticker/page.tsx
@@ -1,5 +1,0 @@
-import AllTickersTable from '@/app/public-equities/ticker/AllTickersTable';
-
-export default function DebugPage() {
-  return <AllTickersTable />;
-}
