+++ insights-ui/src/app/public-equities/tickers/[tickerKey]/page.tsx
