+++ insights-ui/src/app/public-equities/tickers/create/EditTickerView.tsx
