+++ insights-ui/src/app/public-equities/tickers/create/page.tsx
