+++ insights-ui/src/app/public-equities/tickers/edit/[tickerKey]/page.tsx
