+++ insights-ui/src/types/public-equity/ticker.ts
@@ -1,28 +1,5 @@
-export interface Ticker {
-  tickerKey: string;
-  createdAt: Date;
-  updatedAt: Date;
-  sector: string;
-  industryGroup: string;
-}
-
-// ✅ Request type for creating a ticker
 export interface CreateTickerRequest {
   tickerKey: string;
-  sector: string;
-  industryGroup: string;
-}
-
-// ✅ Response type for API
-export interface TickerResponse {
-  success: boolean;
-  message?: string;
-  ticker?: Ticker;
-  error?: string;
-}
-
-export interface AllTickersResponse {
-  success: boolean;
-  tickers?: Ticker[];
-  error?: string;
+  sectorId: number;
+  industryGroupId: number;
 }
